Pulse late stage retry + fixes

+++ src/cryptonote_core/service_node_rules.h
@@ -17,9 +17,9 @@ inline constexpr size_t PULSE_QUORUM_ENTROPY_LAG =
         21;  // How many blocks back from the tip of the Blockchain to source entropy for the Pulse
              // quorums.
 
-inline constexpr size_t PULSE_QUORUM_NUM_VALIDATORS = 11;
-inline constexpr size_t PULSE_QUORUM_SIZE = PULSE_QUORUM_NUM_VALIDATORS + 1 /*Leader*/;
-inline constexpr size_t PULSE_BLOCK_REQUIRED_SIGNATURES =
+inline constexpr int PULSE_QUORUM_NUM_VALIDATORS = 11;
+inline constexpr int PULSE_QUORUM_SIZE = PULSE_QUORUM_NUM_VALIDATORS + 1 /*Leader*/;
+inline constexpr int PULSE_BLOCK_REQUIRED_SIGNATURES =
         7;  // A block must have exactly N signatures to be considered properly
 
 static_assert(PULSE_QUORUM_NUM_VALIDATORS >= PULSE_BLOCK_REQUIRED_SIGNATURES);
@@ -31,8 +31,7 @@ static_assert(
 
 constexpr uint16_t pulse_validator_bit_mask() {
     uint16_t result = 0;
-    for (size_t validator_index = 0; validator_index < PULSE_QUORUM_NUM_VALIDATORS;
-         validator_index++)
+    for (int validator_index = 0; validator_index < PULSE_QUORUM_NUM_VALIDATORS; validator_index++)
         result |= 1 << validator_index;
     return result;
 }
